Update border_law.md - Jay Ryan King
Michal,

Other than some typos, I am making a request to delete Article 38. Residence Visas do not expire.

Also, I am recommending that visas should not be prolonged (extended), as applicants should simply apply for another visa and be processed again. Article 39 has been edited to remove the provisions for extending a visa, and also, the provisions for the interim visa were clarified.

The reason that I do not recommend a prolongation of visas is that personal circumstances change fast enough such that after expiry of a temporary visa, the applicant should prove their status in life have remained the same so that they can still justify a visa.

Article 42 - I added a third item. This is referred to as "Safe Third Country" and is necessary to deter non-genuine asylum claimants.

Let me know what you think?

JK